refactor(analysis): rename parserdialog for all importers (#4008)
