+++ analysis/analysers/importers/SourceMonitorImporter/src/main/kotlin/de/maibornwolff/codecharta/importer/sourcemonitor/Dialog.kt
@@ -1,4 +1,4 @@
-package de.maibornwolff.codecharta.analysis.importer.csv
+package de.maibornwolff.codecharta.analysis.importer.sourcemonitor
 
 import com.varabyte.kotter.runtime.RunScope
 import com.varabyte.kotter.runtime.Session
@@ -9,10 +9,10 @@ import de.maibornwolff.codecharta.dialogProvider.myPromptDefaultFileFolderInput
 import de.maibornwolff.codecharta.dialogProvider.myPromptInput
 import de.maibornwolff.codecharta.serialization.FileExtension
 
-class ParserDialog {
+class Dialog {
     companion object : ParserDialogInterface {
         override fun collectParserArgs(session: Session): List<String> {
-            val inputFileName: String = session.myPromptDefaultFileFolderInput(
+            val inputFileName = session.myPromptDefaultFileFolderInput(
                 inputType = InputType.FILE,
                 fileExtensionList = listOf(FileExtension.CSV),
                 onInputReady = testCallback()
@@ -24,26 +24,7 @@ class ParserDialog {
                 onInputReady = testCallback()
             )
 
-            val pathColumnName: String = session.myPromptInput(
-                message = "What is the name of the path column name?",
-                hint = "path",
-                allowEmptyInput = false,
-                onInputReady = testCallback()
-            )
-
-            val delimiter: String = session.myPromptInput(
-                message = "Which column delimiter is used in the CSV file?",
-                hint = ",",
-                onInputReady = testCallback()
-            )
-
-            val pathSeparator: String = session.myPromptInput(
-                message = "Which path separator is used in the path names?",
-                hint = "/",
-                onInputReady = testCallback()
-            )
-
-            val isCompressed = outputFileName.isEmpty() || session.myPromptConfirm(
+            val isCompressed = (outputFileName.isEmpty()) || session.myPromptConfirm(
                 message = "Do you want to compress the output file?",
                 onInputReady = testCallback()
             )
@@ -51,9 +32,6 @@ class ParserDialog {
             return listOfNotNull(
                 inputFileName,
                 "--output-file=$outputFileName",
-                "--path-column-name=$pathColumnName",
-                "--delimiter=$delimiter",
-                "--path-separator=$pathSeparator",
                 if (isCompressed) null else "--not-compressed"
             )
         }
