+++ analysis/analysers/importers/SourceMonitorImporter/src/main/kotlin/de/maibornwolff/codecharta/importer/sourcemonitor/SourceMonitorImporter.kt
@@ -100,7 +100,7 @@ class SourceMonitorImporter(
             return MetricNameTranslator(replacementMap.toMap(), prefix)
         }
 
-    override fun getDialog(): ParserDialogInterface = ParserDialog
+    override fun getDialog(): ParserDialogInterface = Dialog
 
     override fun isApplicable(resourceToBeParsed: String): Boolean {
         return false
