+++ analysis/analysers/importers/TokeiImporter/src/main/kotlin/de/maibornwolff/codecharta/importer/tokeiimporter/Dialog.kt
@@ -1,4 +1,4 @@
-package de.maibornwolff.codecharta.analysis.importer.codemaat
+package de.maibornwolff.codecharta.analysis.importer.tokeiimporter
 
 import com.varabyte.kotter.runtime.RunScope
 import com.varabyte.kotter.runtime.Session
@@ -9,12 +9,12 @@ import de.maibornwolff.codecharta.dialogProvider.myPromptDefaultFileFolderInput
 import de.maibornwolff.codecharta.dialogProvider.myPromptInput
 import de.maibornwolff.codecharta.serialization.FileExtension
 
-class ParserDialog {
+class Dialog {
     companion object : ParserDialogInterface {
         override fun collectParserArgs(session: Session): List<String> {
             val inputFileName: String = session.myPromptDefaultFileFolderInput(
-                inputType = InputType.FILE,
-                fileExtensionList = listOf(FileExtension.CSV),
+                InputType.FILE,
+                listOf(FileExtension.JSON),
                 onInputReady = testCallback()
             )
 
@@ -25,16 +25,32 @@ class ParserDialog {
                 onInputReady = testCallback()
             )
 
-            val isCompressed =
-                (outputFileName.isEmpty()) ||
-                    session.myPromptConfirm(
-                        message = "Do you want to compress the output file?",
-                        onInputReady = testCallback()
-                    )
+            val rootName = session.myPromptInput(
+                message = "Which root folder was specified when executing tokei?",
+                hint = ".",
+                allowEmptyInput = false,
+                onInputReady = testCallback()
+            )
+
+            var pathSeparator = session.myPromptInput(
+                message = "Which path separator is used? Leave empty for auto-detection",
+                hint = "",
+                allowEmptyInput = true,
+                onInputReady = testCallback()
+            )
+
+            if (pathSeparator == "\\") pathSeparator = "\\\\"
+
+            val isCompressed = (outputFileName.isEmpty()) || session.myPromptConfirm(
+                message = "Do you want to compress the output file?",
+                onInputReady = testCallback()
+            )
 
             return listOfNotNull(
                 inputFileName,
                 "--output-file=$outputFileName",
+                "--root-name=$rootName",
+                "--path-separator=$pathSeparator",
                 if (isCompressed) null else "--not-compressed"
             )
         }
