+++ analysis/analysers/importers/TokeiImporter/src/main/kotlin/de/maibornwolff/codecharta/importer/tokeiimporter/TokeiImporter.kt
@@ -148,7 +148,7 @@ class TokeiImporter(
         if (pathSeparator == "\\\\") this.pathSeparator = "\\"
     }
 
-    override fun getDialog(): ParserDialogInterface = ParserDialog
+    override fun getDialog(): ParserDialogInterface = Dialog
 
     override fun isApplicable(resourceToBeParsed: String): Boolean {
         return false
